model/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.1
gunicorn==20.0.4
pandas==1.0.1
scikit-learn==0.23.2
dash==1.9.0
dash-auth==1.3.2
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-bootstrap-components==0.10.3
requests==2.22.0
SQLAlchemy==1.3.13
Flask-SQLAlchemy==2.4.1
tqdm==4.45.0
mysql-connector-python>=8.0
boto3==1.16.9
scipy==1.4.1
pytorch-lightning
haversine
geopy
dipy
pyproj